Improve codebase
