+++ packages/strapi-admin/admin/src/hooks/useUsersForm/tests/init.test.js
@@ -1,6 +1,6 @@
 import init from '../init';
 
-describe('ADMIN | CONTAINERS | ProfilePage | init', () => {
+describe('ADMIN | HOOKS | useUsersForm | init', () => {
   it('should return the initialState', () => {
     const initialState = {
       test: true,
