+++ packages/strapi-admin/admin/src/hooks/useUsersForm/tests/reducer.test.js
@@ -1,6 +1,6 @@
 import { reducer } from '../reducer';
 
-describe('ADMIN | CONTAINERS | ProfilePage | reducer', () => {
+describe('ADMIN | HOOKS | useUsersForm | reducer', () => {
   describe('DEFAULT_ACTION', () => {
     it('should return the initialState', () => {
       const initialState = {
