+++ packages/strapi-admin/admin/src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as checkFormValidity } from './checkFormValidity';
+export { default as formatAPIErrors } from './formatAPIErrors';
